Abilities/Battle Info
Changed /add_ooch and /add_item to allow autocomplete based on name and ID.
Updated /add_ooch to have a more well put together result message
Added a new button to the Select an Action part of the battles, Show Battle Info, which displays info about both Oochamon on the field as well as general battle information.
Fixed a bug where stats and switch in abilities weren't activated when switching in an oochamon with the Switch button.
Deleted the old map editor from the git repo files