Tested/fixed abilities, battle image changes, and /play in thread functionality
Changed some text in the battle text to be more clear.
Fixed a bug where abilities that changed type effectiveness on moves weren't doing that.
Tested some more abilities to ensure they worked.
Made /play be able to be run inside of a thread, so that it doesn't have to make a new thread constantly.
Made status effects on Oochamon be reset upon leaving and entering battles.
Made the starting battle image display the player Oochamons HP