Evolutions/Rewards Text
Added the ability to evolve an Oochamon in the party menu with a button
Fixed some text in the move selector in the party menu.
Fixed a crash when running /menu when the menu is open.
Fixed a bug where you could not select a move to add to an Oochamon if an Oochamon did not have 4 moves already.
Made the level up/evolution text be part of the Rewards embed.
Fixed some rewards text typos/formatting.