Events and things
Properly made an introduction cutscene run with /setup, and had it transition to a proper setup for the game.
TODO: STILL NEEDS TO DISPLAY THE OOCHAMON INTRODUCTION PICTURE
Made numerous good changes to the event system, to better handle non-NPC events, and condensed some event modes.
Fixed NPCs not showing up and collisions not working properly.