More tweaks
New:
Added functionality for all the status effect items.
Added /lookup item, with basic info so far about the item.
Added functionality for the Repulsor in the main menu.
Added a special message at the top of the battle messages if you have or have not caught the Oochamon you encountered.

Changes:
Changed /teleport to use up emergency teleporters, rather than letting you use it for free.
Changed potions to no longer be usable on dead Oochamon in battle (though they can't be used on other Oochamon on your team, so this doesn't really do much actually.)
Changed where text is placed when an enemy is sending out a new oochamon and triggers an ability to be in a more sensible place.

Fixes:
Fixed bug where the end_of_round screen took double the length it should take to appear.