Menu button/Prism fixes
New:
Added a (currently non-functional) multiplayer button to the menu, replacing the Oochabox button, as well as 2 submenus for it.

Changes:
Changed the prism pulses to be based on how close you were to catching tthe Oochamon.
Moved the remaining functions from func_battle.js to func_battle_new.js, and renamed it back to func_battle.js.

Fixes:
Fixed bug where the playspace message would not be deleted upon starting a battle.