Multiplayer and menu rebuilds
New:
Added /battle with many bugs!

Changes:
Rebuilt the Oochadex menu to use pages, to allow for any number of oochamon (and look cleaner)
Removed the seen number from Oochadex.
Changed the shop image to be above all the text, to make it a bit cleaner.
Changed Multiplayer button to Map button.