Beta Update v3.4 (not ready for release)
New:
Added a special message for winning a battle, to better display who won the battle at the end and separate out end of battle winnings a bit more.
Added the ability to cancel a battle invite through a new button.

Changes:
Changed the Heatwave weather effect to deal 5% of an Oochamon's HP per turn, rather than 10%.
Changed /play to avoid giving you an encounter if ran while you are standing on wild grass.

Fixes:
POTENTIALLY, MAYBE, HOPEFULLY FOR REAL THIS TIME fixed an issue where battle messages would linger on occasion when they should be deleted. PLEASE tell me if this doesn't end up being the case, and what happened in the battle.
Fixed a major bug that would cause battles to "double up" when using /play to force quit a battle.

Just needs to have the box in combat stuff setup. The next push will have the full changelog.