Minor tweaks and function improvements
Changes:
Changed the evolution footer on ooch_info_embed to display ??? for the "evolves into" part if you have not caught the next evolution of the Oochamon.
Changed the tame value stuff to be in it's own function rather than be duplicated code.
Changed the end of prompt code to be in it's own function, to prevent 8 cases of duplicated code (lol)
Updated the stat display message to look a bit nicer.

Fixes:
Fixed a few crashes I noted in #to-do-list